+++ CHANGES.rst
@@ -14,9 +14,11 @@ Unreleased
 * Use ``executemany()`` instead of ``execute()`` in the SQLite storage.
   Makes updating feeds (excluding network calls) 5-10% faster. (`#144`_)
 * In the web app, redirect to the feed's page after adding a feed. (`#119`_)
+* In the web app, show highlighted search result snippets. (`#122`_)
 
 .. _#144: https://github.com/lemon24/reader/issues/144
 .. _#119: https://github.com/lemon24/reader/issues/119
+.. _#122: https://github.com/lemon24/reader/issues/122
 
 
 Version 0.21